DMDB

+++ db/migrate/20210615040733_create_entries.rb
@@ -1,7 +1,7 @@
 class CreateEntries < ActiveRecord::Migration[5.2]
   def change
     create_table :entries do |t|
-      t.references :user, type: :integer, foreign_key: true
+      t.references :user, type: :bigint, foreign_key: true
       t.references :room, type: :integer, foreign_key: true
 
       t.timestamps
